refactor(settings): extract shortcuts cheat-sheet tab to its own file
